% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_regression}
\alias{linear_regression}
\title{linear_regression}
\usage{
linear_regression(X, y)
}
\arguments{
\item{X}{input X, it should be a data.frame object with/without colnames.}

\item{y}{input y, this is the respond variable.}
}
\value{
it will return a list with model coefficients, t statistics and pvalue of every coefficient, and F-stat for the overall model. It will also return R^2 and adjusted R^2, just like everything lm function will return.
}
\description{
This function will do almost exactly the same thing as the famous lm function.
To be specific, It will compute linear regression with the given data.
The data can have factor type and our function will automatically transform it by one-hot encoding.
It will also compute feature siginificance through t-test and an overall F-test just like the lm function.
}
\examples{
library(MASS)
X=Boston[,-ncol(Boston)]
y=Boston[,ncol(Boston)]
linear_regression(X,y)
In fact, using object other than a data.frame will raise an error.
}
