% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_predict.R
\name{regression_predict}
\alias{regression_predict}
\title{regression_predict}
\usage{
regression_predict(regression_object, X)
}
\arguments{
\item{regression_object}{this should be a result computed by linear_regression()}

\item{X}{new_data you want to predict}
}
\value{
it will return a list with model coefficients, t statistics and pvalue of every coefficient, and F-stat for the overall model. It will also return R^2 and adjusted R^2, just like everything lm function will return.
}
\description{
This function mimic the predict function. It takes input of a linear_regression object, designed in the same package.
And it wll use the coefficients in the object to predict by calculating yhat given the other input of a new data.
This function will check the input of your new dataframe object.
If one of the feature used in model training is not appear in the new data, then this function will not compute, and raise an error.
}
\examples{
library(MASS)
X=Boston[,-ncol(Boston)]
y=Boston[,ncol(Boston)]
reg_model = linear_regression(X,y)
yhat = regression_predict(reg_model, X)
#If the new data contains more than necessary columns that are not used, it will be fine.
#'@export
}
