% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_ols.R
\name{simple_ols}
\alias{simple_ols}
\title{simple_ols}
\usage{
simple_ols(X, y)
}
\arguments{
\item{X}{input X, it should be a matrix.}

\item{y}{input y, this is the respond variable.}
}
\value{
it will return coefficient of beta, but it won't compute anything other than beta, such as R^2.
}
\description{
Compute the OLS linear regression for a given X and y. This is the most simple case of ordinary least square. In fact, this function will not perform any error checking because it's used as a internal function of least_square.
}
\examples{
simple_ols(matrix(c(1,3,7,4,6,3),3,2), c(3,1,2))
simple_ols(data.frame(c(1,3,7,4,6,3),3,2) , c(3,1,2))
In fact, using the data.frame object will raise an error.
}
