% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{simple_ols}
\alias{simple_ols}
\title{Compute the OLS linear regression for a given X and y}
\usage{
simple_ols(X, y)
}
\arguments{
\item{X}{input X, it should be a data.frame, otherwise the function will not compute}

\item{y}{input y, this is the respond variable.}
}
\value{
it will return a list object which contains coefficient of beta, and the other important summary statistics of the model
}
\description{
Compute the OLS linear regression for a given X and y
}
